Move Context into its own module
